materializations(refactor): move SQL materializations `client` into a separate file
The materialize-sql `client` abstraction has picked up more responsibility over time, and with some
coming changes to the `Apply` process it will start to encapsulate even more. This commit is a pure
refactoring that moves the `client` for SQL materializations into its own separate file to make
things a little more manageable.
