+++ materialize-motherduck/driver.go
@@ -5,22 +5,15 @@ import (
 	stdsql "database/sql"
 	"encoding/base64"
 	"encoding/json"
-	"errors"
 	"fmt"
-	"net/http"
-	"path"
 	"strings"
-	"time"
 
-	"github.com/aws/aws-sdk-go-v2/aws"
 	awsConfig "github.com/aws/aws-sdk-go-v2/config"
 	"github.com/aws/aws-sdk-go-v2/credentials"
 	"github.com/aws/aws-sdk-go-v2/service/s3"
-	awsHttp "github.com/aws/smithy-go/transport/http"
 	sql "github.com/estuary/connectors/materialize-sql"
 	pf "github.com/estuary/flow/go/protocols/flow"
 	pm "github.com/estuary/flow/go/protocols/materialize"
-	"github.com/google/uuid"
 	log "github.com/sirupsen/logrus"
 	"go.gazette.dev/core/consumer/protocol"
 
@@ -186,175 +179,6 @@ func newDuckDriver() *sql.Driver {
 			}, nil
 		},
 	}
-}
-
-type client struct {
-	db *stdsql.DB
-}
-
-func (c client) Apply(ctx context.Context, ep *sql.Endpoint, req *pm.Request_Apply, actions sql.ApplyActions, updateSpec sql.MetaSpecsUpdate) (string, error) {
-	cfg := ep.Config.(*config)
-
-	db, err := cfg.db(ctx)
-	if err != nil {
-		return "", err
-	}
-	defer db.Close()
-
-	resolved, err := sql.ResolveActions(ctx, db, actions, duckDialect, cfg.Database)
-	if err != nil {
-		return "", fmt.Errorf("resolving apply actions: %w", err)
-	}
-
-	statements := []string{}
-	for _, tc := range resolved.CreateTables {
-		statements = append(statements, tc.TableCreateSql)
-	}
-
-	for _, ta := range resolved.AlterTables {
-		// Duckdb only supports a single ALTER TABLE operation per statement.
-		for _, col := range ta.AddColumns {
-			statements = append(statements, fmt.Sprintf(
-				"ALTER TABLE %s ADD COLUMN %s %s;",
-				ta.Identifier,
-				col.Identifier,
-				col.NullableDDL,
-			))
-		}
-		for _, col := range ta.DropNotNulls {
-			statements = append(statements, fmt.Sprintf(
-				"ALTER TABLE %s ALTER COLUMN %s DROP NOT NULL;",
-				ta.Identifier,
-				col.Identifier,
-			))
-		}
-	}
-
-	for _, tr := range resolved.ReplaceTables {
-		statements = append(statements, tr.TableReplaceSql)
-	}
-
-	action := strings.Join(append(statements, updateSpec.QueryString), "\n")
-	if req.DryRun {
-		return action, nil
-	}
-
-	if len(resolved.ReplaceTables) > 0 {
-		if err := sql.StdIncrementFence(ctx, db, ep, req.Materialization.Name.String()); err != nil {
-			return "", err
-		}
-	}
-
-	// Running these serially is going to be pretty slow, but it's currently not beneficial to use
-	// parallel requests with MotherDuck/DuckDB, so this is the best we can do.
-	for _, s := range statements {
-		if _, err := db.ExecContext(ctx, s); err != nil {
-			return "", fmt.Errorf("executing statement: %w", err)
-		}
-	}
-
-	// Once all the table actions are done, we can update the stored spec.
-	if _, err := db.ExecContext(ctx, updateSpec.QueryString); err != nil {
-		return "", fmt.Errorf("executing spec update statement: %w", err)
-	}
-
-	return action, nil
-}
-
-func (c client) PreReqs(ctx context.Context, ep *sql.Endpoint) *sql.PrereqErr {
-	cfg := ep.Config.(*config)
-	errs := &sql.PrereqErr{}
-
-	pingCtx, cancel := context.WithTimeout(ctx, 20*time.Second)
-	defer cancel()
-
-	if db, err := cfg.db(ctx); err != nil {
-		errs.Err(err)
-	} else if err := db.PingContext(pingCtx); err != nil {
-		errs.Err(err)
-	} else {
-		defer db.Close()
-		var c int
-		if err := db.QueryRowContext(ctx, fmt.Sprintf("SELECT 1 FROM information_schema.schemata WHERE schema_name = '%s'", cfg.Schema)).Scan(&c); err != nil {
-			if errors.Is(err, stdsql.ErrNoRows) {
-				errs.Err(fmt.Errorf("schema %q does not exist", cfg.Schema))
-			} else {
-				errs.Err(err)
-			}
-		}
-	}
-
-	s3client, err := cfg.toS3Client(ctx)
-	if err != nil {
-		// This is not caused by invalid S3 credentials, and would most likely be a logic error in
-		// the connector code.
-		errs.Err(err)
-		return errs
-	}
-
-	// Test creating, reading, and deleting an object from the configured bucket and bucket path.
-	testKey := path.Join(cfg.BucketPath, uuid.NewString())
-
-	var awsErr *awsHttp.ResponseError
-	if _, err := s3client.PutObject(ctx, &s3.PutObjectInput{
-		Bucket: aws.String(cfg.Bucket),
-		Key:    aws.String(testKey),
-		Body:   strings.NewReader("testing"),
-	}); err != nil {
-		if errors.As(err, &awsErr) {
-			// Handling for the two most common cases: The bucket doesn't exist, or the bucket does
-			// exist but the configured credentials aren't authorized to write to it.
-			if awsErr.Response.Response.StatusCode == http.StatusNotFound {
-				err = fmt.Errorf("bucket %q does not exist", cfg.Bucket)
-			} else if awsErr.Response.Response.StatusCode == http.StatusForbidden {
-				err = fmt.Errorf("not authorized to write to %q", path.Join(cfg.Bucket, cfg.BucketPath))
-			}
-		}
-		errs.Err(err)
-	} else if _, err := s3client.GetObject(ctx, &s3.GetObjectInput{
-		Bucket: aws.String(cfg.Bucket),
-		Key:    aws.String(testKey),
-	}); err != nil {
-		if errors.As(err, &awsErr) && awsErr.Response.Response.StatusCode == http.StatusForbidden {
-			err = fmt.Errorf("not authorized to read from %q", path.Join(cfg.Bucket, cfg.BucketPath))
-		}
-		errs.Err(err)
-	} else if _, err := s3client.DeleteObject(ctx, &s3.DeleteObjectInput{
-		Bucket: aws.String(cfg.Bucket),
-		Key:    aws.String(testKey),
-	}); err != nil {
-		if errors.As(err, &awsErr) && awsErr.Response.Response.StatusCode == http.StatusForbidden {
-			err = fmt.Errorf("not authorized to delete from %q", path.Join(cfg.Bucket, cfg.BucketPath))
-		}
-		errs.Err(err)
-	}
-
-	return errs
-}
-
-func (c client) FetchSpecAndVersion(ctx context.Context, specs sql.Table, materialization pf.Materialization) (specB64, version string, err error) {
-	err = c.withDB(func(db *stdsql.DB) error {
-		specB64, version, err = sql.StdFetchSpecAndVersion(ctx, db, specs, materialization)
-		return err
-	})
-	return
-}
-
-func (c client) ExecStatements(ctx context.Context, statements []string) error {
-	return c.withDB(func(db *stdsql.DB) error { return sql.StdSQLExecStatements(ctx, db, statements) })
-}
-
-func (c client) InstallFence(ctx context.Context, checkpoints sql.Table, fence sql.Fence) (sql.Fence, error) {
-	var err = c.withDB(func(db *stdsql.DB) error {
-		var err error
-		fence, err = sql.StdInstallFence(ctx, db, checkpoints, fence, base64.StdEncoding.DecodeString)
-		return err
-	})
-	return fence, err
-}
-
-func (c client) withDB(fn func(*stdsql.DB) error) error {
-	return fn(c.db)
 }
 
 type transactor struct {
